+++ assets/icons/_all.js
@@ -1,0 +1,7 @@
+require('./arrow-right.svg');
+require('./facebook.svg');
+require('./home.svg');
+require('./mail.svg');
+require('./pinterest.svg');
+require('./search.svg');
+require('./twitter.svg');