+++ assets/scripts/main.js
@@ -1,5 +1,8 @@
 require('../styles/main.scss');
 require('../styles/editor-style.scss');
 
+// Icons
+require('../icons/_all.js');
+
 // Composants
 require('../../components/_all.js');